python/requirements.txt
Cython>=0.19.2
numpy>=1.8.1,<=1.14.0
scipy>=0.13.3,<=0.17.0
scikit-image>=0.9.3
matplotlib>=1.5.2
ipython>=3.0.0,<=5.4.1
h5py>=2.2.1,<=2.6.0
leveldb>=0.191
networkx>=1.8.1
nose>=1.3.0
pandas>=0.12.0
pydotplus
python-dateutil>=2
python-gflags>=2.0
pyyaml>=3.10
django==1.11.*
Pillow>=3.3.0,<=4.3.0
six>=1.5.2,<=1.10.0
image>=1.5.16